+++ src/components/nav-secondary.tsx
@@ -18,6 +18,8 @@ import {
 } from "@/registry/new-york-v4/ui/sidebar"
 import { Skeleton } from "@/registry/new-york-v4/ui/skeleton"
 import { Switch } from "@/registry/new-york-v4/ui/switch"
+import { useStatsStore } from "@/store/dashboard-stats-store"
+import { useDataTableStore } from "@/store/data-table-store"
 
 export function NavSecondary({
   items,
@@ -32,7 +34,12 @@ export function NavSecondary({
 } & React.ComponentPropsWithoutRef<typeof SidebarGroup>) {
   const { resolvedTheme, setTheme } = useTheme()
   const [mounted, setMounted] = React.useState(false)
-
+  const fetchEndpointMonitors = useDataTableStore(
+    (state) => state.fetchEndpointMonitors,
+  )
+  const fetchDashboardStats = useStatsStore(
+    (state) => state.fetchDashboardStats,
+  )
   React.useEffect(() => {
     setMounted(true)
   }, [])
@@ -42,6 +49,10 @@ export function NavSecondary({
       <SidebarGroupContent>
         <SidebarMenu className="my-4">
           <AddEndpointMonitorDialog
+            onSuccess={async () => {
+              await fetchEndpointMonitors()
+              await fetchDashboardStats()
+            }}
             trigger={
               <SidebarMenuItem className="flex items-center gap-2">
                 <SidebarMenuButton